requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
sqlalchemy>=2.0.0
plotly>=5.18.0
openpyxl>=3.1.0
pytz>=2023.3
python-dateutil>=2.8.0
pytest>=7.4.0
